move assets.js to project.js
