+++ test/e2e/studio/src/assets/projectAssetType/ProjectAssetTypeWebGpuPipelineConfig.test.js
@@ -1,6 +1,6 @@
 import {getPage} from "../../../../shared/browser.js";
 import {runE2eTest} from "../../../../shared/runE2eTest.js";
-import {createAsset, getAssetTreeView} from "../../../shared/assets.js";
+import {createAsset, getAssetTreeView} from "../../../shared/contentWindows/project.js";
 import {setupNewProject} from "../../../shared/project.js";
 
 await runE2eTest({
