+++ test/e2e/studio/src/projectSelector/projectSelector.test.js
@@ -2,7 +2,7 @@ import {assert, assertEquals, assertExists} from "std/testing/asserts.ts";
 import {log} from "../../../shared/log.js";
 import {runE2eTest} from "../../../shared/runE2eTest.js";
 import {click} from "../../../shared/util.js";
-import {createAsset, getAssetTreeView, waitForAssetExists} from "../../shared/assets.js";
+import {createAsset, getAssetTreeView, waitForAssetExists} from "../../shared/contentWindows/project.js";
 import {getMaybeContentWindowConnectionsElement, waitForContentWindowConnectionsElement} from "../../shared/contentWindows/connections.js";
 import {clickContextMenuItem} from "../../shared/contextMenu.js";
 import {openProjectSelector, setupNewProject, waitForProjectOpen, waitForProjectSelector} from "../../shared/project.js";
